refactor: organizing constants and helpers
