+++ packages/core/src/config.ts
@@ -1,8 +1,32 @@
 import fs from 'node:fs';
 import { isAbsolute, join } from 'node:path';
 import {
-  CSS_DIST_DIR,
   DEFAULT_ASSET_PREFIX,
+  RspackChain,
+  color,
+  debounce,
+  fse,
+  getNodeEnv,
+  isObject,
+  logger,
+  upperFirst,
+} from '@rsbuild/shared';
+import type {
+  InspectConfigOptions,
+  NormalizedConfig,
+  NormalizedDevConfig,
+  NormalizedHtmlConfig,
+  NormalizedOutputConfig,
+  NormalizedPerformanceConfig,
+  NormalizedSecurityConfig,
+  NormalizedServerConfig,
+  NormalizedSourceConfig,
+  NormalizedToolsConfig,
+  RsbuildConfig,
+  RsbuildEntry,
+} from '@rsbuild/shared';
+import {
+  CSS_DIST_DIR,
   DEFAULT_DATA_URL_SIZE,
   DEFAULT_DEV_HOST,
   DEFAULT_MOUNT_ID,
@@ -16,27 +40,9 @@ import {
   SERVER_DIST_DIR,
   SERVICE_WORKER_DIST_DIR,
   SVG_DIST_DIR,
+  TS_CONFIG_FILE,
   WASM_DIST_DIR,
-  color,
-  debounce,
-  getNodeEnv,
-  isObject,
-  logger,
-} from '@rsbuild/shared';
-import type {
-  NormalizedConfig,
-  NormalizedDevConfig,
-  NormalizedHtmlConfig,
-  NormalizedOutputConfig,
-  NormalizedPerformanceConfig,
-  NormalizedSecurityConfig,
-  NormalizedServerConfig,
-  NormalizedSourceConfig,
-  NormalizedToolsConfig,
-  RsbuildConfig,
-  RsbuildEntry,
-} from '@rsbuild/shared';
-import { TS_CONFIG_FILE } from './constants';
+} from './constants';
 import { findExists, isFileExists } from './helpers';
 import { mergeRsbuildConfig } from './mergeConfig';
 import { restartDevServer } from './server/restart';
@@ -384,4 +390,75 @@ export async function loadConfig({
     logger.error(`Failed to load file: ${color.dim(configFilePath)}`);
     throw err;
   }
+}
+
+export async function outputInspectConfigFiles({
+  rsbuildConfig,
+  rawRsbuildConfig,
+  bundlerConfigs,
+  inspectOptions,
+  configType,
+}: {
+  configType: string;
+  rsbuildConfig: NormalizedConfig;
+  rawRsbuildConfig: string;
+  bundlerConfigs: string[];
+  inspectOptions: InspectConfigOptions & {
+    outputPath: string;
+  };
+}) {
+  const { outputPath } = inspectOptions;
+
+  const files = [
+    {
+      path: join(outputPath, 'rsbuild.config.mjs'),
+      label: 'Rsbuild Config',
+      content: rawRsbuildConfig,
+    },
+    ...bundlerConfigs.map((content, index) => {
+      const suffix = rsbuildConfig.output.targets[index];
+      const outputFile = `${configType}.config.${suffix}.mjs`;
+      let outputFilePath = join(outputPath, outputFile);
+
+      // if filename is conflict, add a random id to the filename.
+      if (fse.existsSync(outputFilePath)) {
+        outputFilePath = outputFilePath.replace(/\.mjs$/, `.${Date.now()}.mjs`);
+      }
+
+      return {
+        path: outputFilePath,
+        label: `${upperFirst(configType)} Config (${suffix})`,
+        content,
+      };
+    }),
+  ];
+
+  await Promise.all(
+    files.map((item) =>
+      fse.outputFile(item.path, `export default ${item.content}`),
+    ),
+  );
+
+  const fileInfos = files
+    .map(
+      (item) =>
+        `  - ${color.bold(color.yellow(item.label))}: ${color.underline(
+          item.path,
+        )}`,
+    )
+    .join('\n');
+
+  logger.success(
+    `Inspect config succeed, open following files to view the content: \n\n${fileInfos}\n`,
+  );
+}
+
+export async function stringifyConfig(config: unknown, verbose?: boolean) {
+  // webpackChain.toString can be used as a common stringify method
+  const stringify = RspackChain.toString as (
+    config: unknown,
+    options: { verbose?: boolean },
+  ) => string;
+
+  return stringify(config as any, { verbose });
 }
