+++ packages/plugin-react/src/splitChunks.ts
@@ -48,7 +48,6 @@ export const applySplitChunksRule = (
 
     chain.optimization.splitChunks({
       ...currentConfig,
-      // @ts-expect-error Rspack and Webpack uses different cacheGroups type
       cacheGroups: {
         ...(currentConfig as Exclude<SplitChunks, false>).cacheGroups,
         ...createCacheGroups(extraGroups),
