--- packages/shared/src/config.ts
@@ -1,80 +1,0 @@
-import { join } from 'node:path';
-import fse from '../compiled/fs-extra/index.js';
-import { logger } from './logger';
-import type { InspectConfigOptions, NormalizedConfig } from './types';
-import { color, upperFirst } from './utils';
-
-export async function outputInspectConfigFiles({
-  rsbuildConfig,
-  rawRsbuildConfig,
-  bundlerConfigs,
-  inspectOptions,
-  configType,
-}: {
-  configType: string;
-  rsbuildConfig: NormalizedConfig;
-  rawRsbuildConfig: string;
-  bundlerConfigs: string[];
-  inspectOptions: InspectConfigOptions & {
-    outputPath: string;
-  };
-}) {
-  const { outputPath } = inspectOptions;
-
-  const files = [
-    {
-      path: join(outputPath, 'rsbuild.config.mjs'),
-      label: 'Rsbuild Config',
-      content: rawRsbuildConfig,
-    },
-    ...bundlerConfigs.map((content, index) => {
-      const suffix = rsbuildConfig.output.targets[index];
-      const outputFile = `${configType}.config.${suffix}.mjs`;
-      let outputFilePath = join(outputPath, outputFile);
-
-      // if filename is conflict, add a random id to the filename.
-      if (fse.existsSync(outputFilePath)) {
-        outputFilePath = outputFilePath.replace(/\.mjs$/, `.${Date.now()}.mjs`);
-      }
-
-      return {
-        path: outputFilePath,
-        label: `${upperFirst(configType)} Config (${suffix})`,
-        content,
-      };
-    }),
-  ];
-
-  await Promise.all(
-    files.map((item) =>
-      fse.outputFile(item.path, `export default ${item.content}`),
-    ),
-  );
-
-  const fileInfos = files
-    .map(
-      (item) =>
-        `  - ${color.bold(color.yellow(item.label))}: ${color.underline(
-          item.path,
-        )}`,
-    )
-    .join('\n');
-
-  logger.success(
-    `Inspect config succeed, open following files to view the content: \n\n${fileInfos}\n`,
-  );
-}
-
-export async function stringifyConfig(config: unknown, verbose?: boolean) {
-  const { default: RspackChain } = await import(
-    '../compiled/rspack-chain/index.js'
-  );
-
-  // webpackChain.toString can be used as a common stringify method
-  const stringify = RspackChain.toString as (
-    config: unknown,
-    options: { verbose?: boolean },
-  ) => string;
-
-  return stringify(config as any, { verbose });
-}
