+++ packages/shared/src/index.ts
@@ -5,6 +5,5 @@ export * from './getBrowserslist';
 export * from './logger';
 export * from './reduceConfigs';
 export * from './types';
-export * from './config';
 export * from './chain';
 export * from './url';
